linking contacts and users, adding search

+++ src/controllers/Contacts.controller.js
@@ -4,28 +4,61 @@ import { Database } from "../../database.js"
 const database = new Database()
 
 const tableContacts = "contacts"
+const tableUsers = "users"
 
-const AllContacts = (req, res) => {
-  res.json(database.select(tableContacts))
+const AllContactsAndSearch = (req, res) => {
+  const search = req.query.search;
+
+  if(!search){
+    res.json(database.select(tableContacts));
+    return
+  }
+
+  let result = [];
+  database.select(tableContacts).map((contact) => {
+    if(contact.name.toLowerCase().includes(search.toLowerCase())){
+      result.push(contact);
+      return
+    }
+    if(contact.number.toLowerCase().includes(search.toLowerCase())){
+      result.push(contact);
+      return
+    }
+  })
+  
+  res.json(result);  
 }
-const SearchContacts = (req, res) => {
-  res.send(`Busca um contato ${req.params.busca}`);
-}
+
 const AddContact = (req, res) => {
   const {
     id = '',
-    name = '',
-    number = '',
-    imgUrl = ''
+    name,
+    number,
+    imgUrl = '',
+    userId,
   } = req.body;
 
   if (id) { // Caso o id seja informado, é atualizado
+    if (!database.select(tableContacts, id)) {
+      res.status(400).json({ message: 'Contato não encontrado', joke: 'proveded id not found in db' });
+      return
+    }
     database.update(tableContacts, id, {
       name,
       number,
-      imgUrl
-    })
+      imgUrl,
+    });
     res.send({ message: 'Contato atualizado', contact: { id, name, number, imgUrl } });
+    return
+  }
+  if (!userId) { // Caso o userId NÃO seja informado é retornado ERRO
+    res.status(400).json({ message: 'Usuário criador do contato não adicionado', joke: 'userId is required in body' });
+    return
+  }
+
+  const userFromId = database.select(tableUsers, userId);
+  if (!userFromId) { // Rejeita caso o userId não seja encontrado no banco
+    res.status(400).json({ message: 'Usuário criador do contato não encontrado', joke: 'userId not found' });
     return
   }
 
@@ -33,20 +66,27 @@ const AddContact = (req, res) => {
     id: randomUUID(),
     name,
     number,
-    imgUrl
+    imgUrl,
+    owner: userId,
   }
+
   database.insert(tableContacts, user);
-  res.status(201).json({ message: 'Contato adicionado', contact: user});
+  database.update(tableUsers, userFromId.id, { contacts: [...userFromId.contacts, user.id] });
+
+  res.status(201).json({ message: 'Contato adicionado e vinculado a um usuário', contact: user});
 }
 const DeleteContact = (req, res) => {
   const { id } = req.params
+  if (!id) {
+    res.status(400).json({ message: 'Contato não encontrado', joke: 'proveded id not found in params' });
+    return
+  }
   database.delete(tableContacts, id);
   res.send({ message: 'Contato deletado', id});
 }
 
 export {
-  AllContacts,
-  SearchContacts,
+  AllContactsAndSearch,
   AddContact,
-  DeleteContact
+  DeleteContact,
 }
